+++ cmd/containers-storage/main.go
@@ -8,6 +8,7 @@ import (
 	"github.com/containers/storage/internal/opts"
 	"github.com/containers/storage/pkg/mflag"
 	"github.com/containers/storage/pkg/reexec"
+	"github.com/containers/storage/types"
 	"github.com/sirupsen/logrus"
 )
 
@@ -32,7 +33,7 @@ func main() {
 		return
 	}
 
-	options := storage.StoreOptions{}
+	options := types.StoreOptions{}
 	debug := false
 
 	makeFlags := func(command string, eh mflag.ErrorHandling) *mflag.FlagSet {
@@ -67,7 +68,7 @@ func main() {
 	}
 
 	if options.GraphRoot == "" && options.RunRoot == "" && options.GraphDriverName == "" && len(options.GraphDriverOptions) == 0 {
-		options, _ = storage.DefaultStoreOptionsAutoDetectUID()
+		options, _ = types.DefaultStoreOptionsAutoDetectUID()
 	}
 	args := flags.Args()
 	if len(args) < 1 {
